Split surface_service, and undo a baseline I raised without noticing
Four commits ago, decomposing `update_surface` and `_sync_email_schedule` grew
`surface_service.py` from 954 lines to 1025 -- and I regenerated the baseline
over it. That is exactly the move the baseline header forbids, and I did not
catch it because the ratchet passes trivially against a baseline you have just
rewritten. The file is 500 lines now and off the oversized list entirely, and
the ratchet was re-checked against the baseline as it stood *before* that
commit to prove nothing else was quietly raised on the way.

  surface_service.py  1025 -> 500
    surface_email_schedule.py (226)   the polling schedule behind Gmail/Outlook
    surface_telegram_webhook.py (223) registering and revoking one bot's webhook
    surface_consent.py (246)          tenant admin consent and the setup guide

Each takes its own module-level constants with it, which is the other half of
that lesson: `_EMAIL_TRIGGER_EVENT_TYPES` and `_WEBHOOK_RETRY_POLICY` belong
beside the code that reads them, not in whichever file they were declared in
first.

Two test patch-targets moved with their code, and one is now aimed better than
it was: twelve tests patched
`services.surface_service.settings.api_url` to reach a global singleton through
whichever module happened to import it. They patch `app.core.config.settings`
directly now, which is where it actually lives and will not break again the next
time that import moves.

`agent_surfaces` stays at zero complex functions and zero oversized files.
Repo-wide: 26 oversized files -> 24.

5,329 unit tests, 173 surface e2e tests, and `make typecheck-critical` pass.
